+++ packages/eyes-nightwatch/commands/eyesCheck.js
@@ -1,8 +1,5 @@
-const Events = require('events')
-
-module.exports = class EyesCheck extends Events {
+module.exports = class EyesCheck {
   async command(args) {
     await this.client.api.globals.__eyes.check(args)
-    this.emit('complete')
   }
 }