+++ packages/eyes-nightwatch/commands/eyesClose.js
@@ -1,8 +1,5 @@
-const Events = require('events')
-
-module.exports = class EyesClose extends Events {
+module.exports = class EyesClose {
   async command(args) {
     await this.client.api.globals.__eyes.close(args)
-    this.emit('complete')
   }
 }