+++ packages/eyes-nightwatch/commands/eyesOpen.js
@@ -1,24 +1,19 @@
 'use strict'
-const Events = require('events')
-const {Eyes, ConsoleLogHandler, VisualGridRunner, Configuration} = require('../dist')
+const {Eyes, VisualGridRunner} = require('../dist')
 
-module.exports = class EyesOpen extends Events {
+module.exports = class EyesOpen {
   async command(appName, testName, viewportSize) {
     const config = (this.client.options && this.client.options.eyes) || {}
 
     let eyes = this.client.api.globals.__eyes
     if (!eyes) {
+      if (config.enableEyesLogs) {
+        config.logs = {type: 'console'}
+      }
       const runner = config.useVisualGrid ? new VisualGridRunner({testConcurrency: config.concurrency}) : undefined
-      this.client.api.globals.__eyes = eyes = new Eyes(runner)
+      this.client.api.globals.__eyes = eyes = new Eyes(runner, config)
     }
 
-    if (config.enableEyesLogs) {
-      eyes.setLogHandler(new ConsoleLogHandler(true))
-    }
-
-    const configuration = new Configuration(config)
-    eyes.setConfiguration(configuration)
     await eyes.open(this.client.api, appName, testName, viewportSize)
-    this.emit('complete')
   }
 }