+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix return driver from runner
 
 ## 12.21.1 - 2021/6/15